requirements.txt
discord.py
requests
google-genai